Refactor time limit parameter to double in solver algorithms and enhance benchmark metrics
- Changed the time limit parameter type from `unsigned int` to `double` in the `solve` methods of LTMOA, NWMOA, SOPMOA, SOPMOA_bucket, SOPMOA_relaxed algorithms for improved precision.
- Introduced a new `benchmark_metrics.cpp` file to implement benchmark recording functionalities, including metrics tracking and CSV output for frontier and trace data.
- Added a smoke test for benchmark metrics to validate the functionality of the `BenchmarkRecorder` and ensure correct metrics are recorded and outputted.
- Updated the main application to support new command-line options for budget and output paths, enhancing the flexibility of the benchmarking process.

+++ src/algorithms/emoa.cpp
@@ -1,7 +1,7 @@
 #include"algorithms/emoa.h"
 
 template<int N>
-void EMOA<N>::solve(unsigned int time_limit) {
+void EMOA<N>::solve(double time_limit) {
     std::cout << "start EMOA\n";
 
     Label<N>* curr;
